setup
now working as tdlistener
